// libs/nao-http2/src/lib/nao-socket-io.interface.ts
export namespace SocketIOClient {

  /**
   * The base emiter class, used by Socket and Manager
   */
  export interface Emitter {
    /**
     * Adds a listener for a particular event. Calling multiple times will add
     * multiple listeners
     * @param event The event that we're listening for
     * @param fn The function to call when we get the event. Parameters depend on the
     * event in question
     * @return This Emitter
     */
    on( event: string, fn: Function ): Emitter;

    /**
     * @see on( event, fn )
     */
    addEventListener( event: string, fn: Function ): Emitter;

    /**
     * Adds a listener for a particular event that will be invoked
     * a single time before being automatically removed
     * @param event The event that we're listening for
     * @param fn The function to call when we get the event. Parameters depend on
     * the event in question
     * @return This Emitter
     */
    once( event: string, fn: Function ): Emitter;

    /**
     * Removes a listener for a particular type of event. This will either
     * remove a specific listener, or all listeners for this type of event
     * @param event The event that we want to remove the listener of
     * @param fn The function to remove, or null if we want to remove all functions
     * @return This Emitter
     */
    off( event: string, fn?: Function ): Emitter;

    /**
     * @see off( event, fn )
     */
    removeListener( event: string, fn?: Function ): Emitter;

    /**
     * @see off( event, fn )
     */
    removeEventListener( event: string, fn?: Function ): Emitter;

    /**
     * Removes all event listeners on this object
     * @return This Emitter
     */
    removeAllListeners(): Emitter;

    /**
     * Emits 'event' with the given args
     * @param event The event that we want to emit
     * @param args Optional arguments to emit with the event
     * @return Emitter
     */
    emit( event: string, ...args: any[] ): Emitter;

    /**
     * Returns all the callbacks for a particular event
     * @param event The event that we're looking for the callbacks of
     * @return An array of callback Functions, or an empty array if we don't have any
     */
    listeners( event: string ): Function[];

    /**
     * Returns if we have listeners for a particular event
     * @param event The event that we want to check if we've listeners for
     * @return True if we have listeners for this event, false otherwise
     */
    hasListeners( event: string ): boolean;
  }
// ...
  export interface Socket extends Emitter {

    /**
     * The Manager that's controller this socket
     */
    io: SocketIOClient.Manager;

    /**
     * The namespace that this socket is for
     * @default '/'
     */
    nsp: string;

    /**
     * The ID of the socket; matches the server ID and is set when we're connected, and cleared
     * when we're disconnected
     */
    id: string;

    /**
     * Are we currently connected?
     * @default false
     */
    connected: boolean;

    /**
     * Are we currently disconnected?
     * @default true
     */
    disconnected: boolean;

    /**
     * Opens our socket so that it connects. If the 'autoConnect' option for io is
     * true (default), then this is called automatically when the Socket is created
     */
    open(): Socket;

    /**
     * @see open();
     */
    connect(): Socket;

    /**
     * Sends a 'message' event
     * @param args Any optional arguments that we want to send
     * @see emit
     * @return This Socket
     */
    send( ...args: any[] ): Socket;

    /**
     * An override of the base emit. If the event is one of:
     * 	connect
     * 	connect_error
     * 	connect_timeout
     * 	connecting
     * 	disconnect
     * 	error
     * 	reconnect
     * 	reconnect_attempt
     * 	reconnect_failed
     * 	reconnect_error
     * 	reconnecting
     * 	ping
     * 	pong
     * then the event is emitted normally. Otherwise, if we're connected, the
     * event is sent. Otherwise, it's buffered.
     *
     * If the last argument is a function, then it will be called
     * as an 'ack' when the response is received. The parameter(s) of the
     * ack will be whatever data is returned from the event
     * @param event The event that we're emitting
     * @param args Optional arguments to send with the event
     * @return This Socket
     */
    emit( event: string, ...args: any[] ): Socket;

    /**
     * Disconnects the socket manually
     * @return This Socket
     */
    close(): Socket;

    /**
     * @see close()
     */
    disconnect(): Socket;

    /**
     * Sets the compress flag.
     * @param compress If `true`, compresses the sending data
     * @return this Socket
     */
    compress(compress: boolean): Socket;
  }
// ...
  export interface Manager extends Emitter {

    /**
     * All the namespaces currently controlled by this Manager, and the Sockets
     * that we're using to communicate with them
     */
    nsps: { [namespace: string]: Socket };

    /**
     * The connect options that we used when creating this Manager
     */
    opts: SocketIOClient.ConnectOpts;

    /**
     * The state of the Manager. Either 'closed', 'opening', or 'open'
     */
    readyState: string;

    /**
     * The URI that this manager is for (host + port), e.g. 'http://localhost:4000'
     */
    uri: string;

    /**
     * The currently connected sockets
     */
    connecting: Socket[];

    /**
     * If we should auto connect (also used when creating Sockets). Set via the
     * opts object
     */
    autoConnect: boolean;

    /**
     * Gets if we should reconnect automatically
     * @default true
     */
    reconnection(): boolean;

    /**
     * Sets if we should reconnect automatically
     * @param v True if we should reconnect automatically, false otherwise
     * @default true
     * @return This Manager
     */
    reconnection( v: boolean ): Manager;

    /**
     * Gets the number of reconnection attempts we should try before giving up
     * @default Infinity
     */
    reconnectionAttempts(): number;

    /**
     * Sets the number of reconnection attempts we should try before giving up
     * @param v The number of attempts we should do before giving up
     * @default Infinity
     * @return This Manager
     */
    reconnectionAttempts( v: number ): Manager;

    /**
     * Gets the delay in milliseconds between each reconnection attempt
     * @default 1000
     */
    reconnectionDelay(): number;

    /**
     * Sets the delay in milliseconds between each reconnection attempt
     * @param v The delay in milliseconds
     * @default 1000
     * @return This Manager
     */
    reconnectionDelay( v: number ): Manager;

    /**
     * Gets the max reconnection delay in milliseconds between each reconnection
     * attempt
     * @default 5000
     */
    reconnectionDelayMax(): number;

    /**
     * Sets the max reconnection delay in milliseconds between each reconnection
     * attempt
     * @param v The max reconnection dleay in milliseconds
     * @return This Manager
     */
    reconnectionDelayMax( v: number ): Manager;

    /**
     * Gets the randomisation factor used in the exponential backoff jitter
     * when reconnecting
     * @default 0.5
     */
    randomizationFactor(): number;

    /**
     * Sets the randomisation factor used in the exponential backoff jitter
     * when reconnecting
     * @param The reconnection randomisation factor
     * @default 0.5
     * @return This Manager
     */
    randomizationFactor( v: number ): Manager;

    /**
     * Gets the timeout in milliseconds for our connection attempts
     * @default 20000
     */
    timeout(): number;

    /**
     * Sets the timeout in milliseconds for our connection attempts
     * @param The connection timeout milliseconds
     * @return This Manager
     */
    timeout(v: number): Manager;

    /**
     * Sets the current transport socket and opens our connection
     * @param fn An optional callback to call when our socket has either opened, or
     * failed. It can take one optional parameter of type Error
     * @return This Manager
     */
    open( fn?: (err?: any) => void ): Manager;

    /**
     * @see open( fn );
     */
    connect( fn?: (err?: any) => void ): Manager;

    /**
     * Creates a new Socket for the given namespace
     * @param nsp The namespace that this Socket is for
     * @return A new Socket, or if one has already been created for this namespace,
     * an existing one
     */
    socket( nsp: string ): Socket;
  }
// ...
}